backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
pydantic==2.11.3
pydantic-settings==2.8.1
qdrant-client==1.14.2
sentence-transformers==3.4.1
google-genai==1.13.0
httpx==0.28.1
beautifulsoup4==4.13.3
readability-lxml==0.8.1
orjson==3.10.16
Pillow==10.4.0
rapidfuzz==3.13.0
unidecode==1.3.8
python-dotenv==1.1.0
